+++ packages/client/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.20-beta
 - Stop fetching client data on `setActive(false)`
+- Remove `moment` as dependency
 
 # 0.0.19-beta
 - Add rollup